Moved the namespaces into a namespaces folder.
